Update product and ingredient controller and service

+++ src/main/java/fr/polytech/melusine/models/dtos/requests/UserRegistrationRequest.java
@@ -21,7 +21,7 @@ public class UserRegistrationRequest {
     @NotNull
     private Section section;
 
-    private double credit;
+    private long credit;
 
     private boolean isMembership;
